pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["Juan Carlos Gomez <“juancarlosgomez@cecotec.es”>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
djangorestframework = "^3.13.1"
twilio = "^7.13.0"
pip = "^22.2.2"
install = "^1.3.5"
django-middleware-global-request = "^0.3.0"
Faker = "^14.2.0"
django-configurations = "^2.4"
django-debug-toolbar = "^3.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
